80 excel header rule test cases

+++ app/src/EditGroups.tsx
@@ -122,13 +122,18 @@ function handleRemoveGroup(
   groups: Map<string, Group>,
   groupName: string
 ): Map<string, Group> {
-  groupRuleService.deleteGroup(groupName).then(response => {
-    if (response.status !== 200) {
-      alert("deleteGroup fail with status: " + response.status);
-    }
-  });
-  groups.delete(groupName);
-  return groups;
+  const group = groups.get(groupName);
+  if (group) {
+    groupRuleService.deleteGroup(groupName).then(response => {
+      if (response.status !== 200) {
+        alert("deleteGroup fail with status: " + response.status);
+      }
+    });
+    groups.delete(groupName);
+    return new Map(groups);
+  } else {
+    return groups
+  }
 }
 
 function handleAddGroupRule(
@@ -171,6 +176,7 @@ function handleUpdateGroupName(
       }
     });
     groups.delete(input.oldGroupName);
+    group.name = input.newGroupName;
     return new Map([...groups.set(input.newGroupName, group).entries()].sort());
   } else {
     return groups;
